test: add cases about unquoted workbook prefixes
Workbook names in unquoted reference prefixes are currently permitted to
include spaces and double-quote characters. Rules about this are not specified
in any published grammar, I think, but other parsers permit it too, and because
of the square brackets it does not cause any lexical ambiguity. Seems harmless.

This just adds test cases to make it clear and to guard against regressions.