Switch documentation for from/to funcs in a1/rc
The README's documentation on the `from` and `to` functions in the `a1`
module appears to be the wrong way round. For example, the docs for
`a1.to()` says that it takes a `columnString` argument, and you use the
function to

    Parse a simple string reference to an A1 range into a range object
    (see above). Will accept A1, A2, A:A, or 1:1.

However, it _actually_ takes `range` argument that's expected to be an
object, not a string, and it will return a string.

The opposite is true for the `a1.from()` function: it takes a string and
returns an object, but the README says you use it to

    Stringify a range object (see above) into A1 syntax.

Example usage:

    import { a1 } from "@borgar/fx";

    a1.to({ top: 9, bottom: 9, left: 2, right: 2 })
    // Returns "C10"

    a1.from("C10")
    // Returns a range object:
    //
    //   {
    //     top: 9,
    //     left: 2,
    //     bottom: 9,
    //     right: 2,
    //     '$top': false,
    //     '$left': false,
    //     '$bottom': false,
    //     '$right': false
    //   }

The same is true for the `to` and `from` functions in the `rc` module.

I think all that's happened here is that the documentation for the two
`to` functions has been included in the README under the `from`
functions, and vice versa.