types: convey location guarantee in return types
When `tokenize` is called with `withLocation: true` (or without `withLocation`,
since `true` is the default), make the return type convey that the `loc`
property of returned tokens will not be `undefined`.

Add the same type parameter on `parse` although its return type is currently
just `object` (so it's there if/when that return type gets fleshed out).

Type-parametrize `addTokenMeta` so that this token type information is not lost
when `tokenize` results are passed through it.