Workaround: apply loc type guarantee with patch
The types are generated by jsdoc, which does not support nontrivial type
expressions and [does not seem likely to do so in the future][1]. Using
different tooling would be a hassle.

So just apply these minimal type changes with a patch when building types.

If future changes cause the jsdoc type output to conflict with this patch,
then that will not go unnoticed as the build:types run-script will fail.
If/when that happens, one can just manually edit `dist/fx.d.ts` and regenerate
the patch with `git diff dist/fx.d.ts > convey-loc-guarantee-in-types.patch`.

[1]: https://github.com/jsdoc/jsdoc/issues/1917